格式化代码

+++ app/src/main/java/net/kaaass/zerotierfix/util/IPPacketUtils.java
@@ -1,7 +1,9 @@
 package net.kaaass.zerotierfix.util;
 
 import android.util.Log;
+
 import androidx.core.view.MotionEventCompat;
+
 import java.net.InetAddress;
 import java.net.UnknownHostException;
 
@@ -67,7 +69,7 @@ public class IPPacketUtils {
     public static long calculateChecksum(byte[] bArr, long j, int i, int i2) {
         int i3 = i2 - i;
         while (i3 > 1) {
-            j += (long) ((65280 & (bArr[i] << 8)) | (bArr[i + 1] & 255));
+            j += (65280 & (bArr[i] << 8)) | (bArr[i + 1] & 255);
             if ((-65536 & j) > 0) {
                 j = (j & 65535) + 1;
             }
@@ -75,7 +77,7 @@ public class IPPacketUtils {
             i3 -= 2;
         }
         if (i3 > 0) {
-            j += (long) ((bArr[i] << 8) & MotionEventCompat.ACTION_POINTER_INDEX_MASK);
+            j += (bArr[i] << 8) & MotionEventCompat.ACTION_POINTER_INDEX_MASK;
             if ((j & -65536) > 0) {
                 j = (j & 65535) + 1;
             }
